crud

+++ app/controllers/restaurants_controller.rb
@@ -54,7 +54,7 @@ class RestaurantsController < ApplicationController
     @cuisines = []
     @restaurant = Restaurant.find(params[:id])
     @states = []  
-  @type = ['1','2']
+    @type = ['1','2']
     Cuisine.all.each do |x|
       if @restaurant.cuisines.include?(x) then puts x else @cuisines << x end 
     end
@@ -67,7 +67,7 @@ class RestaurantsController < ApplicationController
       
     if @restaurant.update_attributes(restaurant_params)
       edit_nests
-      redirect_to edit_user_restaurant_path, notice: 'Sweet Edit Bro' 
+      redirect_to edit_user_restaurant_path, notice: 'Thanks' 
     else 
     render json: @restaurant.errors 
     end
@@ -109,31 +109,33 @@ class RestaurantsController < ApplicationController
           end
         end
         #edit employee roles
-        params[:restaurant][:aware_employees_attributes].each do |employee|
-          this_emp = AwareEmployee.where(name:employee[1][:name], restaurant_id: @restaurant.id).first
-          next if this_emp == nil
-          if employee[1][:role_ids] != nil
-             employee[1][:role_ids].each do |role|  
-               if role == ''
-                 RestaurantRole.where(aware_employee_id:this_emp.id).each do |record|
-                   record.destroy
+        if params[:restaurant][:aware_employees_attributes] != nil
+          params[:restaurant][:aware_employees_attributes].each do |employee|
+            this_emp = AwareEmployee.where(name:employee[1][:name], restaurant_id: @restaurant.id).first
+            next if this_emp == nil
+            if employee[1][:role_ids] != nil
+               employee[1][:role_ids].each do |role|  
+                 if role == ''
+                   RestaurantRole.where(aware_employee_id:this_emp.id).each do |record|
+                     record.destroy
+                   end
+                 else
+                  RestaurantRole.create(aware_employee_id:this_emp.id, role_id:role)
+                  AwareEmployee.where(name:this_emp.name, restaurant_id: @restaurant.id).each do |x|
+                    if x != this_emp then x.destroy end
+                  end
                  end
-               else
-                RestaurantRole.create(aware_employee_id:this_emp.id, role_id:role)
-                AwareEmployee.where(name:this_emp.name, restaurant_id: @restaurant.id).each do |x|
-                  if x != this_emp then x.destroy end
-                end
                end
-             end
-          else
-            employee[1][:restaurant_roles_attributes].first[1][:role_id].each do |role|
-              if role !='' then RestaurantRole.create(aware_employee_id:this_emp.id, role_id:role) end
-            end
-            AwareEmployee.where(name:this_emp.name, restaurant_id: @restaurant.id).each do |x|
-                  if x != this_emp then x.destroy end
+            else
+              employee[1][:restaurant_roles_attributes].first[1][:role_id].each do |role|
+                if role !='' then RestaurantRole.create(aware_employee_id:this_emp.id, role_id:role) end
+              end
+              AwareEmployee.where(name:this_emp.name, restaurant_id: @restaurant.id).each do |x|
+                    if x != this_emp then x.destroy end
+              end
             end
           end
-        end  
+        end 
     end
     def save_nests(this_restaurant)
       # Save Cuisine Nest
@@ -141,22 +143,17 @@ class RestaurantsController < ApplicationController
         if cuisine != '' then TypeOfCuisine.create(restaurant_id: this_restaurant.id, cuisine_id: cuisine ) end
       end
       # End Cuisine
-
       # Save Neighborhood Nest
         params[:restaurant][:areas_attributes].first[1][:neighborhood_id].each do |hood|
           if hood != '' then Area.create(restaurant_id: this_restaurant.id, neighborhood_id: hood )end
         end
-
       # End Neighborhood  
-     
       # Save Employee Roles
       #   iterate through aware_employees
-  
          params[:restaurant][:aware_employees_attributes].each do |employee|
           #find saved emplpyee with the correct name AT the current restaurant
            saved_employee = AwareEmployee.where(name: employee[1][:name], restaurant_id: this_restaurant.id).first
         #   #iterate through restaurant roles and save 
-          
            employee[1][:restaurant_roles_attributes].first[1][:role_id].each do |role|
              if role != '' then RestaurantRole.create(aware_employee_id: saved_employee.id, role_id: role ) end
             end
